Include custom finder method in DataTables->find() options array
When $finder is set as second argument of the find function, it can
cause errors if user fails to set it and instead sets the $options array
as the second argument. This change makes it so the custom finder
method, if used, is included in the options array instead of as second
argument, improving usability in addition to following CakePHP best
practices. Also updates readme to reflect availability of this feature.